Don't forbid the X-XSRF-TOKEN header for annotation API http requests.
Authentication is handled by the X-Annotator-Auth-Token header,
so the X-XSRF-TOKEN is not necessary, however if the client has
visited the server app with the same browser, then he received
a cookie there, so he will keep transmitting this header,
with any request sent to the annotation storage - even with
cross-site requests, coming from other applications.

However, those requests were rejected, because of the presence
of the header. Allowing it does no harm, and adds no function,
simply permits the browser to talk to us, even if it happens to have
this cookie incidentally.